requirements.txt
matplotlib==3.2.2
numpy==1.16.4
Pillow~=6.2.0
torch==1.7.1
torchvision==0.8.2
torchsummary~=1.5.1